requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
asgiref==3.9.1; python_version >= '3.9'
astroid==3.3.11; python_full_version >= '3.9.0'
attrs==25.3.0; python_version >= '3.8'
certifi==2025.8.3; python_version >= '3.7'
charset-normalizer==3.4.3; python_version >= '3.7'
deprecation==2.1.0
dill==0.4.0; python_version >= '3.8'
django==5.2.5; python_version >= '3.10'
django-cors-headers==4.7.0; python_version >= '3.9'
django-crontab==0.7.1
django-filter==25.1; python_version >= '3.9'
django-jazzmin==3.0.1
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.0; python_version >= '3.9'
docopt==0.6.2
drf-nested-routers==0.94.2; python_version >= '3.8'
drf-spectacular==0.28.0; python_version >= '3.7'
factory-boy==3.3.3; python_version >= '3.8'
faker==37.6.0; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
h11==0.16.0; python_version >= '3.8'
h2==4.2.0; python_version >= '3.9'
hpack==4.1.0; python_version >= '3.9'
httpcore==1.0.9; python_version >= '3.8'
httpx[http2]==0.28.1; python_version >= '3.8'
hyperframe==6.1.0; python_version >= '3.9'
idna==3.10; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
isort==6.0.1; python_full_version >= '3.9.0'
jsonschema==4.25.1; python_version >= '3.9'
jsonschema-specifications==2025.4.1; python_version >= '3.9'
mailjet-rest==1.5.1; python_version >= '3.10'
mccabe==0.7.0; python_version >= '3.6'
mysqlclient==2.2.7; python_version >= '3.8'
num2word==1.0.1
num2words==0.5.14
numpy==2.3.1; python_version >= '3.11'
packaging==25.0; python_version >= '3.8'
pandas==2.3.1; python_version >= '3.9'
pillow==11.3.0; python_version >= '3.9'
platformdirs==4.3.8; python_version >= '3.9'
postgrest==1.1.1; python_version >= '3.9' and python_version < '4.0'
psycopg2-binary==2.9.10; python_version >= '3.8'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pyjwt==2.9.0; python_version >= '3.8'
pylint==3.3.8; python_full_version >= '3.9.0'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-decouple==3.8
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
realtime==2.5.3; python_version >= '3.9' and python_version < '4.0'
referencing==0.36.2; python_version >= '3.9'
reportlab==4.4.2; python_version >= '3.7' and python_version < '4'
requests==2.32.5; python_version >= '3.9'
rpds-py==0.27.1; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlparse==0.5.3; python_version >= '3.8'
storage3==0.12.0; python_version >= '3.9' and python_version < '4.0'
strenum==0.4.15
supabase==2.16.0; python_version >= '3.9' and python_version < '4.0'
tomlkit==0.13.3; python_version >= '3.8'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
uritemplate==4.2.0; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
websockets==15.0.1; python_version >= '3.9'
whitenoise==6.9.0; python_version >= '3.9'
